refactor: 인수테스트 시나리오 기반으로 리팩토링
